all: introduce a syscall flag that controls support checking
For some syscalls (openat, mount) syzkaller always checks const string
arguments. E.g. it can check whether the file that is intended to be
opened by openat is present on the VM.

This works fine for almost all cases and is used as a means to prevent
fuzzing of non-supported features. However, it prevents us from refering
to those files/devices that will only be created either at the time of
fuzzing or during syz-executor initialization.

Introduce an `always_supported` option to flag syscalls which must
bypass such checks. Flag two binderfs syscalls.