all: add the `rerun` call property
To be able to collide specific syscalls more precisely, we need to
repeat the process many times.

Introduce the `rerun` call property, which instructs `syz-executor` to
repeat the call the specified number of times. The intended use is:

call1() (rerun: 100, async)
call2() (rerun: 100)

For now, assign rerun values randomly to consecutive pairs of calls,
where the first one is async.