syz-fuzzer: split WorkSmash
It was observed that some smash tasks may take up to 30 minutes. This is
not particularly efficient, as in the meanwhile there can appear other
candidate or triage assignments. Split smash into multiple subtasks so
that the average wait time for the higher priority tasks stays as low as
possible.