syz-fuzzer: split work queue into global and per-group ones
It turns out that syz-fuzzer tends to spread similar activities to
almost all its procs and they fuzz similar subsystems at the same time
and it is not very efficient. For example, when all the procs become
busy fuzzing a subsystem with blocking calls, the whole VM becomes
underloaded and stay idle most of the time.

Prevent this by splitting the single work queue into multiple ones.

syz-testbed experiments show that this version finds more coverage and
20-30% more crashes at any given moment of execution, when executed with
no prior corpus.