all: refactor discussion processing code
Now that too much logic seems to be duplicated in tools/syz-lore and
dahsboard/app/discussion.go, it's time to refactor the code.

Factor out the code that decides whether an incoming email message
should start a new thread or be appended to the previous one.

Restructure pkg/email/lore so that email processing matches the
one-by-one approach of reporting_email.go.