all: support cross-tree fix bisection
If tree origin assessment code has identified that the bug is not
reproducible in a tree from which we merge/cherry-pick commits, use fix
bisection to identify that commit. This can e.g. be used to find fixing
commits that were not backported from Linux kernel mainline into LTS
branches.

In case of bisection errors, re-do such jobs every 30 days.

Store candidate fixes in the Bug structure. Later the information will be
displayed on the web dashboard and reported to bug discussions.