pkg/compiler: support if[expr] attributes
The expression may either include integers/consts or reference other
fields in the structure via value[field1:field2:field3].

The fields on this path must all belong to structures and must not have
any if conditions themselves.

For unions, mandate that the last field has no conditions (it will be
the default one).

For structs, convert conditional fields into fields of a union type of
the following form:

anonymous_union [
 value T (if[expression])
 void void
]