pkg/fuzzer: redistribute hint and smash iterations
Do fewer smashes and hint iterations for new corpus items. This should
not be a big problem since if any new coverage is found, we would
eventually repeat the same process with only a slightly different
program.

Also, to somehow compensate for this, start new hint jobs for existing
corpus items once in a while.